refactor some of busting files to the new architecture
